+++ src/Game/blobNetwork/draw.ts
@@ -6,13 +6,10 @@ import {
   ENTRANCE_RADIUS_X,
   ENTRANCE_RADIUS_Y,
   CONNECTION_RADIUS_PERCENT,
-  NODE_RADIUS_X,
-  NODE_RADIUS_Y,
 } from 'game/paramaters';
-import { Ellipse, Point } from 'game/types';
-import { drawCircle } from 'game/lib/draw';
+import { Ellipse } from 'game/types';
 
-import { Node, Connection, NodeMap } from './types';
+import { Connection } from './types';
 
 const nodeColor = hexToRGB(blobQueenColor);
 
@@ -121,104 +118,4 @@ export function drawNodeConnectionRadius(
   ctx.stroke();
   ctx.closePath();
   ctx.setLineDash([]);
-}
-
-function drawNodeConnectionRadii(ctx: CanvasRenderingContext2D, nodes: Node[]) {
-  nodes.forEach((node) => drawNodeConnectionRadius(ctx, node));
-}
-
-function drawConnectionPoint(
-  ctx: CanvasRenderingContext2D,
-  { x, y }: Point,
-  color = blobQueenColor
-) {
-  ctx.beginPath();
-  drawCircle(ctx, x, y, 14, color);
-  ctx.strokeStyle = 'black';
-  ctx.stroke();
-  ctx.closePath();
-}
-
-function drawPendingConnectionLine(
-  ctx: CanvasRenderingContext2D,
-  start: Point,
-  end: Point,
-  color = 'black'
-) {
-  ctx.setLineDash([5, 15]);
-
-  ctx.beginPath();
-  ctx.moveTo(start.x, start.y);
-  ctx.lineTo(end.x, end.y);
-  ctx.strokeStyle = color;
-  ctx.stroke();
-  ctx.closePath();
-  ctx.setLineDash([]);
-}
-
-export function drawChoosingStart(
-  ctx: CanvasRenderingContext2D,
-  nodes: Node[],
-  startPoint?: Point
-) {
-  drawNodeConnectionRadii(ctx, nodes);
-
-  if (startPoint) {
-    drawConnectionPoint(ctx, startPoint);
-  }
-}
-
-function makeColor(isOnNode?: boolean, isValid?: boolean) {
-  if (!isValid) return 'red';
-  return isOnNode ? blobQueenColor : 'black';
-}
-
-export function drawChoosingEnd(
-  ctx: CanvasRenderingContext2D,
-  nodes: Node[],
-  start?: Point,
-  end?: Point,
-  isOnNode?: boolean,
-  isValid?: boolean
-) {
-  if (!start) return;
-  const color = makeColor(isOnNode, isValid);
-
-  drawNodeConnectionRadii(ctx, nodes);
-  drawConnectionPoint(ctx, start);
-
-  if (end) {
-    drawConnectionPoint(ctx, end, color);
-    drawPendingConnectionLine(ctx, start, end, color);
-  }
-}
-
-export function drawAdjustingEnd(
-  ctx: CanvasRenderingContext2D,
-  nodes: Node[],
-  start?: Point,
-  end?: Point,
-  endNodeCentre?: Point
-) {
-  if (!start || !end) return;
-
-  drawNodeConnectionRadii(ctx, nodes);
-
-  drawConnectionPoint(ctx, start);
-  drawConnectionPoint(ctx, end);
-
-  drawPendingConnectionLine(ctx, start, end);
-
-  if (endNodeCentre) {
-    drawNode(ctx, {
-      centre: endNodeCentre,
-      radiusX: NODE_RADIUS_X,
-      radiusY: NODE_RADIUS_Y,
-    });
-    drawNodeConnectionRadius(ctx, {
-      centre: endNodeCentre,
-      radiusX: NODE_RADIUS_X,
-      radiusY: NODE_RADIUS_Y,
-    });
-  }
 }
